vuex setup and user details stored to state successfully
