+++ resources/js/app.js
@@ -5,13 +5,11 @@ import Vue from 'vue';
 import App from './components/App.vue';
 import router from './router.js';
 import store from './store';
-// import httpClient from './api/httpClient';
 
 const app = new Vue({
     el: '#app',
     components: { App },
     router,
     store,
-    // httpClient
 
 });
