+++ resources/js/store.js
@@ -4,44 +4,46 @@ import Vuex from 'vuex';
 Vue.use(Vuex);
 
 
-// import the api endpoints
-import {fetchUser } from "./api/users.api";
-
 /*
 *   store/modules/users.module.js
 */
 
 const state = {
-    users: []
-}
+    users: {}
+};
 
 const getters = {
-    loginUser(state) {
+    getUser(){
         return state.users;
     }
-}
 
-const actions = {
-    async fetchUsers({ commit }) {
-        try {
-            const response = await loginUser();
-            commit('SET_USERS', response.data);
-        } catch (error) {
-            // handle the error here
-        }    
-    }
-}
+};
 
 const mutations = {
     SET_USERS(state, data) {
         state.users = data;
     }
-}
+};
 
-export default {
-    namespaced: true,
+const actions = {
+    // async fetchUsers({ commit }) {
+    //     try {
+    //         const response = await loginUser();
+    //         commit('SET_USERS', response.data);
+    //     } catch (error) {
+    //         // handle the error here
+    //     }    
+    // }
+
+    SET_USER(context, user) {
+        context.commit("SET_USER", user);
+    },
+};
+
+
+export default new Vuex.Store( {
     state,
     getters,
-    actions,
-    mutations
-}
+    mutations,
+    actions
+})
